requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
Django==5.0.3
django-ninja==1.1.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
